+++ .github/workflows/build-apps-web.yml
@@ -10,7 +10,7 @@ on:
       - '.github/workflows/build-apps-web.yml'
       - 'apps/web/**'
       - 'packages/db/**'
-      - 'packages/trpc/**'
+      - 'packages/api/**'
       - 'packages/web/**'
       - 'packages/config/eslint/**'
       - 'packages/config/tsconfig/**'
@@ -22,7 +22,7 @@ on:
       - '.github/workflows/build-apps-web.yml'
       - 'apps/web/**'
       - 'packages/db/**'
-      - 'packages/trpc/**'
+      - 'packages/api/**'
       - 'packages/web/**'
       - 'packages/config/eslint/**'
       - 'packages/config/tsconfig/**'
