+++ apps/web/package.json
@@ -11,6 +11,7 @@
     "test": "jest"
   },
   "dependencies": {
+    "@acme/api": "workspace:*",
     "@acme/theme": "workspace:*",
     "@acme/web-forms": "workspace:*",
     "@acme/web-locales": "workspace:*",
@@ -30,8 +31,7 @@
     "react-hook-form": "^7.38.0",
     "react-i18next": "^11.18.6",
     "react-query": "^3.39.2",
-    "superjson": "^1.10.0",
-    "trpc": "workspace:*"
+    "superjson": "^1.10.0"
   },
   "devDependencies": {
     "@acme/eslint-config": "workspace:*",
