+++ packages/api/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "trpc",
+  "name": "@acme/api",
   "version": "0.0.0",
   "license": "MIT",
   "main": "dist/index.js",
